Blackduck issue induced by terraform-plugin-docs
terraform-plugin-docs has indirect dependency of yaml.v2,
which induce blackduck security issue.
Now instead of using the tools.go we add the command in the main.go.
This way the provider can just pull the dependency and
generate the docs when needed and doesn't need to pull these into the go.mod.